Изменена страничная навигация, смена языка интерфейса
