+++ messages/de.json
@@ -21,9 +21,11 @@
   "LatestEntries": {
     "dateNotSet": "Datum nicht festgelegt"
   },
-  "LocaleSwitcher": {
-    "label": "Change language",
-    "locale": "{locale, select, de {🇩🇪 Deutsch} en {🇺🇸 English} other {Unknown}}"
+  "Footer": {
+    "localeSwitcher": {
+      "label": "Interface language",
+      "locale": "{locale, select, de {🇩🇪 Deutsch} en {🇺🇸 English} other {Unknown}}"
+    }
   },
   "NotFoundPage": {
     "title": "Page not found",
