+++ messages/es.json
@@ -21,9 +21,11 @@
   "LatestEntries": {
     "dateNotSet": "Fecha no establecida"
   },
-  "LocaleSwitcher": {
-    "label": "Change language",
-    "locale": "{locale, select, de {🇩🇪 Deutsch} en {🇺🇸 English} other {Unknown}}"
+  "Footer": {
+    "localeSwitcher": {
+      "label": "Interface language",
+      "locale": "{locale, select, de {🇩🇪 Deutsch} en {🇺🇸 English} other {Unknown}}"
+    }
   },
   "NotFoundPage": {
     "title": "Page not found",
