+++ src/components/LocaleSwitcher.tsx
@@ -1,19 +1,14 @@
 'use client'
 import {useLocale} from 'next-intl';
 import {locales} from '@/config';
-import {Dropdown} from '@mui/base/Dropdown';
-import {Menu} from '@mui/base/Menu';
-import {MenuButton} from '@mui/base/MenuButton';
 import React, {useTransition} from "react";
-import {MenuItem} from '@mui/base/MenuItem';
 import {usePathname, useRouter} from "@/navigation";
 import {useSearchParams} from "next/navigation";
 import {Suspense} from "react";
-import clsx from "clsx";
-import {className} from "postcss-selector-parser";
+import LocaleSwitcherSelect from "@/components/LocaleSwitcherSelect";
 
 
-export default function LocaleSwitcher() {
+export default function LocaleSwitcher({label}: { label: string }) {
 
     const router = useRouter();
     const [isPending, startTransition] = useTransition();
@@ -24,7 +19,7 @@ export default function LocaleSwitcher() {
     const createHandleMenuClick = (cur: string) => {
         const params = new URLSearchParams(searchParams);
         return () => {
-             startTransition(() => {
+            startTransition(() => {
                 router.push(
                     // @ts-expect-error -- TypeScript will validate that only known `params`
                     // are used in combination with a given `pathname`. Since the two will
@@ -38,23 +33,13 @@ export default function LocaleSwitcher() {
 
     return (
         <Suspense>
-            <Dropdown>
-                <MenuButton
-                    className="p-3 sm:text-base/7 text-sm">{locale.toUpperCase()
-                }</MenuButton>
-                <Menu className="bg-svoddWhite-400 dark:bg-svoddBlack-400 border rounded-xl p-7">
-                    {locales.map((cur) => (
-                        <MenuItem
-                            key={cur} onClick={createHandleMenuClick(cur)}
-                            className={clsx('cursor-pointer', className, {
-                                'font-bold': locale === cur
-                            })}
-                        >
-                            {cur.toUpperCase()}
-                        </MenuItem>
-                    ))}
-                </Menu>
-            </Dropdown>
+            <LocaleSwitcherSelect defaultValue={locale} label={label}>
+                {locales.map((cur) => (
+                    <option key={cur} value={cur}>
+                        {cur}
+                    </option>
+                ))}
+            </LocaleSwitcherSelect>
         </Suspense>
     );
 }
