+++ src/components/LocaleSwitcherSelect.tsx
@@ -43,14 +43,14 @@ export default function LocaleSwitcherSelect({
         >
             <p className="sr-only">{label}</p>
             <select
-                className="inline-flex appearance-none bg-transparent py-3 pl-2 pr-6"
+                className="inline-flex appearance-none bg-transparent py-3 pl-2 pr-8 border-0"
                 defaultValue={defaultValue}
                 disabled={isPending}
                 onChange={onSelectChange}
             >
                 {children}
             </select>
-            <span className="pointer-events-none absolute right-2 top-[8px]">⌄</span>
+            {/*<span className="pointer-events-none absolute right-2 top-[8px]">⌄</span>*/}
         </label>
     );
 }
