+++ src/ui/global.scss
@@ -30,6 +30,7 @@ article, .article, .search-panel {
     @apply inline-flex items-center px-3.5 py-0.5 rounded-sm text-sm font-medium;
     cursor: pointer;
   }
+
   mark {
     @apply inline-block pr-1 pl-1 py-0;
     @apply bg-svoddWhite-100 text-svoddBlack-100 dark:bg-svoddGray-400 dark:text-svoddWhite-200;
@@ -43,11 +44,28 @@ article, .article, .search-panel {
   .MuiPaginationItem-ellipsis {
     @apply dark:text-svoddWhite-200 dark:bg-transparent;
   }
+
   .MuiPaginationItem-page, .MuiPaginationItem-previousNext {
     @apply text-svoddRed-400 dark:text-svoddRed-300 dark:border-svoddGray-300 border-pageLight-border;
   }
+
   .Mui-selected {
-    @apply dark:text-svoddWhite-200 text-white dark:border-svoddRed-700 dark:bg-svoddBlack-200 border-svoddRed-900 !bg-svoddRed-800;
+    @apply dark:text-svoddWhite-200 text-white dark:border-svoddRed-700 dark:bg-svoddBlack-200 border-svoddRed-900 bg-svoddRed-800;
     @apply bg-svoddWhite-400 dark:bg-svoddRed-500;
   }
 }
+
+.pagination {
+  .detachable {
+    margin: 1rem 0;
+  }
+
+  .fixed-bottom {
+    position: fixed;
+    right: 0;
+    bottom: 0;
+    left: 0;
+    //z-index: 1029;
+  }
+}
+
